+++ packages/camera/camera_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.11.0+11
+
+* Adds hardware burst capture support for Android using Camera2.
+
 ## 0.10.10+11
 
 * Updates examples to use the new RadioGroup API instead of deprecated Radio parameters.